requirements.txt
Django==2.0.6
django-disqus==0.5
django-tagging==0.4.6
mock==2.0.0
pbr==5.1.1
Pillow==5.3.0
pytz==2018.4
six==1.11.0
